+++ packages/cnpj-dv/src/cnpj_dv/__init__.py
@@ -1,17 +1,27 @@
-from .cnpj_check_digits import CnpjCheckDigits
-from .exceptions import (
-    CnpjCheckDigitsCalculationError,
-    CnpjCheckDigitsError,
-    CnpjInvalidLengthError,
-    CnpjTypeError,
+from .cnpj_check_digits import (
+    CNPJ_MAX_LENGTH,
+    CNPJ_MIN_LENGTH,
+    CnpjCheckDigits,
 )
+from .exceptions import (
+    CnpjCheckDigitsException,
+    CnpjCheckDigitsInputInvalidException,
+    CnpjCheckDigitsInputLengthException,
+    CnpjCheckDigitsInputTypeError,
+    CnpjCheckDigitsTypeError,
+)
+from .types import CnpjInput
 
 __all__ = [
+    "CNPJ_MAX_LENGTH",
+    "CNPJ_MIN_LENGTH",
     "CnpjCheckDigits",
-    "CnpjCheckDigitsCalculationError",
-    "CnpjCheckDigitsError",
-    "CnpjInvalidLengthError",
-    "CnpjTypeError",
+    "CnpjCheckDigitsException",
+    "CnpjCheckDigitsInputInvalidException",
+    "CnpjCheckDigitsInputLengthException",
+    "CnpjCheckDigitsInputTypeError",
+    "CnpjCheckDigitsTypeError",
+    "CnpjInput",
 ]
 
 __version__ = "0.0.0"